#35 - refactor:  CategoryType enum 적용을 위한 리팩토링
특정 카테고리에 대한 정보에 접근할 때 String으로 접근하던 것을 CategoryType enum을 통해 접근하도록 함

+++ generator-simulator/src/main/java/com/simpaylog/generatorsimulator/dto/Trade.java
@@ -1,6 +1,8 @@
 package com.simpaylog.generatorsimulator.dto;
 
+import java.math.BigDecimal;
+
 public record Trade(
         String tradeName,
-        int cost
+        BigDecimal cost
 ) {}
